[Templates] Error for C++ Empty activity without generateLayout
When there is a layout file, we find a TextView with
a specific resource ID and set it with the result of
calling stringFromJNI().
In this CL, if there there is no layout file generated,
just use a simple System.out.println() to demo how
to make a call to a native method.

Test: UI tests in PROJECT_WIZARD group, TemplateTest and
manually test Java/Kotlin combinations
Test-strategy: ui.test.groups=PROJECT_WIZARD
fixes: 78203315
Change-Id: Ia633d1062d33f42306eb045cf08e3cb92ff85606