Fix multi module test
The new MultiModule test has tests that rely on implementation
dependencies working fine. Unfortunately, this is only supported
in V2 as V1 inheritly cannot support it (app has to generate
everything for dependencies).

This CL fixes the test to avoid using implementation when
V2 is disabled.

It also excludes non-androidX tests as we don't update them
anymore (they are snapshots of ToT when we moved to AndroidX
and kept for backwards compatibility verification).

Bug: 77539932
Test: DataBindingExternalArtifactDependencyConnectedTest
Change-Id: Ic962cf217f9ab040f7050abd96af9ef6a42e7c22