Switch cxx integration tests to CMake 3.10
Bug: 80429739
Test: This is a test only change

This also switches to ninja 1.8.3 which allows for more filename
segments in paths. CMake 3.10.2 isn't yet part of the SDK so this
change uses cmake.dir to access the CMake instead. Once CMake is
part of the SDK we could switch back to using
locate-in-SDK-by-version logic but it's not necessary.

Change-Id: Id1c5a583a284040c76b5d8734eec4465ae55ab53