Optimize hashCode for repeated calls to parent
Prior to this patch, hashCode() runs in O(n) worst-case time. That means
that an algorithm that iterates over the parent folders by calling
getParent() O(n) times would degenerate to O(n^2) if it calls
hashCode on each iteration. (Note that this would happen for Strings too
but we can do better.)

However, we can make the hashCode run in O(1) in this situation by
changing the hashing function. If we hash each segment normally but then
add the hashes together with a simple sum (or any other reversible
function), we can compute the hashcode for the parent folder from the
hashcode of any of its children by just subtracting off the hash of the
last segment.

This patch implements such a hashing function. Note that - like any other
change to a hashing function - a possible consequence of this change may
be more hash collisions. This hasn't been observed in practice and I
don't consider it likely for the sort of data that normally shows up in
system paths, but if a lot of hash collisions are observed for PathString
instances, it should be safe to revert this change.

Test: PathStringTest.testHashCodeForParent
Change-Id: Ib4e5da4dd673ec9744881697011a475ac624d4a9