Fix MemoryTest flakiness
Move tags initialization outside the loop so allocation and deallocation
events don't have to be returned in the right order for deallocations to
be counted. Also re-organize code to query JVMTI data from endTimestamp
of the previous call.

With the fix this test flakes 1/200 and the flake is related to
connection failures. Will investigate and address that in a follow-up
CLs.

Bug: 80501379
Test: N/A
Change-Id: I7eb8174c0c677e32e6d920e590ebd2d70ed27561