Convert Jacoco transform to task
This CL introduces a new internal artifact
type, for Jacoco instrumented classes.

A helper method has been added to TransformManager,
to help with migration from transforms to tasks. In this
particular case, the method is used to consume streams
that would be consumed by the Jacoco transforms. Generated
artifacts are added back to the available streams, so the
rest of the transform pipeline works as it used to.

BUG: 111156401
Test: changed existing
Change-Id: Iff0e6e86dbbbcb6b038dba350be3accbea33a2db