Use absolute path of input to compute output path in transforms
This CL changes how we compute output path for transform
inputs. It moves us away from using the content name to
compute the output location, and it uses the file absolute
path instead.

Reason for this change is the correctness of incremental builds.
E.g. there could an input that contains a directory content
which name changes across runs. This means that the output location
for such input will change between two runs. So in an
incremental build, we will fail to remove the old output,
and that can result in issues for downstream transforms.

BUG: 110818090
Test: DexArchiveBuilderTransformTest, DexMergerTransformTest
Change-Id: I60b3d479deee9defbc1f043869798c0f64bf5c6d