Split java-lib-model-builder imls
Turns out provided depencies end up on the classpath when running tests
through JPS and our gradle-api brings in SLF4J classes that conflict
with other logging classes on the IDE classpath. The gradle-api jar that
IDEA comes with (the "Gradle" library) was apparently stripped to not
include those, so we'll use it in the tools/idea project.

Test: manual
Bug: N/A
Change-Id: I9801f9f4c0c182e2ea9e49f1eb79dd7ef01f24e1