Adds a "swap" command to the installer.
Creates a swap command on the installer, which takes a SwapRequest proto
via stdin.

The command copies the agent and instrumentation binaries into the
data directory of the app specified by the SwapRequest, then invokes the
hot-swap by passing the SwapRequest to the agent.

Note that no output from the agent is read (yet).
That behavior is left for a follow-up CL.

The installer will not copy the binaries if it sees existing binaries
in the folder, unless it is passed a 1 on the command line.

This is temporary behavior to allow us to easily test manually; a better
protocol for updates is needed.

Test: Manual. Create a SwapRequest with the proto_tool, then invoke the
install_to_device script and check logcat.
Bug: 111804699
Change-Id: I99be842068b7f126f83856ccab4e7214aeca6c6b