CMake version pinning: relax CMake version matching
Before this CL "3.2" in build.gradle would not match "3.2.1-rc1"
and now it will. The new semantic is that the version must match
exactly up-to the value in build.gradle. After that, the highest
version on the path is taken. So if "3.2.0" is specified in
build.gradle then 3.2.1-rc1 won't be matched.

Bug: 110693527
Test: Added unittests to cover

Change-Id: I7c8c13431b7adbf51a65768e315ee2aed488c182