Added reflection cache for injection points.
Calculating tasks injection points using reflection will now
be cached. The cache is a static field as it only needs to
be refreshed when tasks implementation changes themselves.

Cache contains a list of input and output injection points
indexed per task type.

When tasks are configured, the injection point is used to
allocate BuildableArtifacts or look them up. When tasks
are finalized before execution, injection will be
performed [no new behavior].

Some TODOs and clean up are left for a subsequent CL to
not complicate review further.

Test: existing and new ones.
BUG: N/A
Change-Id: I3b75bcdb0e6bf01da499ae8811be859c060f2ef0