Allow simpleperf to debug system_server
system_server process is managed by the system and as such can't be run
with run-as. We need to handle this special case in order to support
profiling this process using simpleperf. That can be achieved by preventing
the usage of --app flag when profiling system_server and making sure to run
simpleperf record as root. Naturally, in order to kill simpleperf and finish
recording, we also need to send the SIGTERM signal as root.

Bug: 112562093
Test: simpleperf_manager_test.cc and simpleperf_test.cc

Change-Id: I0bb47bc4bdb6121db817e157f8114c176e1c3afb